ref: fix some mypy issues in snuba tests (#53458)
<!-- Describe your PR here. -->

+++ src/sentry/testutils/cases.py
@@ -1673,7 +1673,7 @@ class MetricsEnhancedPerformanceTestCase(BaseMetricsLayerTestCase, TestCase):
 
     def store_transaction_metric(
         self,
-        value: List[int] | int,
+        value: List[float] | float,
         metric: str = "transaction.duration",
         internal_metric: Optional[str] = None,
         entity: Optional[str] = None,
